Remove config ScriptToRun
And inline in into Config object
